bowyer watson

+++ src/delauney-triangulation/bowyerWatson.h
@@ -3,4 +3,4 @@
 #include "../vizualizationModule.h"
 #include "geomAdditionalFunc.h"
 #pragma once
-void triangulation(std::vector<Point2d>& points,std::vector<Triangle>& triangulation);
+void triangulation(std::vector<Point2f>& points,std::vector<Triangle>& triangulation);
